ui: select menu items by Cypress through ids
In Cypress it's difficult to choose a specific item in the menu if you
don't know in which language you are. Each element can be in French,
English, etc.
To facilitate this work we can add `id=` on each element on the menu.
With a specific word. This way you don't need to know in which language
you are.

* Adds Id on all menus ('My Account' menu, languages menu, help menu,
logout, etc.) for testing purposes.

Co-Authored-by: Olivier DOSSMANN <git@dossmann.net>
Co-Authored-by: Bertrand Zuchuat <bertrand.zuchuat@rero.ch>